Remove the pre-pended apostrophe from reference numbers.
Reference numbers from the MeS API always have apostrophes appended
to them. The MeS documentation is inconsistent in its examples, showing
apostrophes pre-pended to some and none pre-pended to others. In the MeS
reporting front-end, apostrophes never appear next to the reference numbers,
and the apostrophe is never mentioned in the documentation.

Further, the reference number is documented as a 23-character field. With
the apostrophe, this field consistently comes to 24.

This leads me to believe that this is sloppiness on MeS' end. This commit
makes it the default to remove the apostrophe from reference numbers. Folks
using this API who *do* want to keep the apostrophe for any reason -- personal
preference, interfacing with legacy records -- can get the reference number
unchanged from records by passing in `:unchanged => true` thusly:

    reference_number = record.reference_number(:unchanged => true)

There were also a few shoulds missing from the tests. Adding these, we
saw a couple of failures, but none of them were bugs -- we just had the
wrong values to expect.